Add coverage and node_modules to .gitignore
A case could be made that eventually we want
include the `coverage` directory for CI/CD.
However, right now it is just bukly. We can
start checking it in later if we so desire.

For node_modules, we really shouldn't commit
this. The package-lock.json will handle
ensuring we have the proper dependencies
installed. However, if you really want to
pad your github lines added/deleted, start
checking it in.