Fix basic app.test.jsx test
The basic test in app.test.jsx was failing due to it not having a
store. The error the test was giving was complaining that it should
be wrapped in a <Provider. I decided to humor it and also gave it
a mocked store (which was also a necessary part of resolving the
error).