Refactor Launch.jsx into pure component
Launch didn't have any state to track, so there was no reason for
it to be a fully defined class. The linter was also complaining
about this. So I refactored it into a pure component, disabled the
camelCase requirement for the file, and propgated it's changed
call signature as needed.