Add settings to eslint for webpack aliases
Eslint didn't recognize webpack aliases for imports which made it
sad. Firstly I added `eslint-import-resolver-webpack` as a dev
dependency so that we don't depend on somone having it globaly
installed via npm on their machine. Secondly, I've added settings
in .eslintrc.js so that eslint knows where to look for webpack
resolution definitions.