Simply Layout.jsx return value
Eslint was complaining about the unnecessary {} after the => in
Layout.jsx because it's technically unnecessary. One solution is
remove the {} and the `return()` just leaving the jsx after the
arrow. However, I prefer some sort of encapsulation for readability
so I went with () which eslint doesn't have a problem with.