Update app render test to care about a full mount
When we previously updated the app.test.jsx render test, we updated
it to use a shallow rendering. This actually didn't ensure the full
rendering of the app without crashing. It only ensured that Routes
was being instantiated without crashing. However, we want to test
that all the sub components also don't crash. To do so we must use
`mount` instead of `shallow`. Additionally we need to provide a
proper store state because of this.