Refactor Launch and have Launches handle list keys
Previous the Launch component always assumed it would be in a list.
This may not always be the case, so it makes sense for it not to use
<li> as it's outer most wrapper. The real reason we did this though
is because Launch is now a pure component, which means Launches now
needs to control the `key` for elements in it's list that are
dynamically generated.