Convert Launches view into a function component
I plan on using react hooks for the new functionality as I prefer
that methodology for maintaining lifecycle changes / managing state.
Most of the updates will be from the `Launches` view on down, so it
felt right to convert it so my state management is consistent. I
could update the provider store to use useContext instead of using
redux, and I still might, but for the time being I've decided to
leave it be and see how I after actually implementing what you've
asked me to do.