Remove launches from redux store
The launch collection data is specific to the launches page, there
isn't an entirely great reason to store it in redux. If it were that
we were using the same information in a bunch of components it may
be worthwhile to have it in the global store, but for the time being
this is unnecessary. If we do ever bring it back into the global
store, we should also ensure that we have a way to invalidate and
refresh the data.